Removing old database code

+++ hackertracker/src/main/java/com/shortstack/hackertracker/Event/FavoriteEvent.kt
@@ -1,3 +1,3 @@
 package com.shortstack.hackertracker.Event
 
-class FavoriteEvent(val item : Int)
+class FavoriteEvent(val item : Long)
